Rendering: Refactored vulkan abstraction
We splitted the files into multiple directories and we abstracted
command buffer, command pool, fence and semaphores into files.

The handles are now stored in ref counted wrappers as temporary
solution. We should prefer to store them raw.

The renderer should not use the raw vulkan types, we need to move
the vulkan functionality to seperate classes.