Rendering: Changed way of passing wrappers
We decided to go with the first approach to pass the handles around,
but we managed to remove the dependency of using Rc or RefCell.

Instead of relying on the Drop trait, which would require cloning
and storing of the vulkan handles in the wrapper, we decided to
manually clean up the vulkan handles via cleanup function.

The cleanup function requires the passing of the handles from the
context itself and the wrapper is required to pass the handle down.

Because we removed the requirement to store the handles in the wrapper
we need to pass the handles to the renderer every time we call a
function from the renderer. This could lead to a significant performance
downgrade, which we should investigate.