Rendering: Added class specific abstractions back
We removed the class specific abstractions, because we handled with raw
vulkan handles.

In the last patch we changed the way of passing handles and decided to
use the wrappers. The patch caused that we can add the class specific
abstractions back in to simplify the way of using them.