Rendering: Use window time for manipulating scene
Before this patch the flash light from the blue clear color and the
rotating of the model was based on the current frame. The current frame
is much faster on GPU's with MAILBOX, which would increase the refresh
rate of the engine.

Now we use the get_time function from glfw to get the time since start
which does not change based on the refresh rate. (This maybe has to be
multiplied with the delta time)