Engine: Made some 'use' changes
Replaced a lot of repeating prefixes of modules with a use at the top of
the file. This will make the code more readable and takes less space in
the screen.

This change does not affect the rendering module, because the module
will be refactored later on with a big change.